backEnd/requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pymysql>=1.0.2
dbutils>=3.0.2
pymysqlreplication>=0.27
matlabengine>=9.11.0